backend/requirements.txt
# ============================================================================
# RPA ORCHESTRATOR - REQUIREMENTS
# ============================================================================

# FastAPI e Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic e Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# SQLModel e Database
sqlmodel==0.0.14
sqlalchemy==2.0.25
alembic==1.13.1

# PostgreSQL Drivers
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0

# HTTP Client (para testes e integrações)
httpx==0.26.0

# Validation & Utilities
python-dateutil==2.8.2

# ============================================================================
# DEVELOPMENT & TESTING (opcional, comentado por padrão)
# ============================================================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.0.0
# black==24.1.1
# flake8==7.0.0
# mypy==1.8.0
# isort==5.13.2alembic==1.13.1
aiosqlite==0.18.0